% Generated by roxygen2 (4.0.1): do not edit by hand
\name{A_efficiency}
\alias{A_efficiency}
\title{A-efficiency}
\usage{
A_efficiency(Design)
}
\arguments{
\item{Design}{A block design data frame such as returned by \code{\link{blocks}}}
}
\description{
\code{A_efficiency} A-efficiency of an existing design
}
\details{
A-efficiencies for nested blocks design in a data frame with a
 column of block factor levels for the total blocks in each
 nested blocks stratum and a final column containing the treatment
 factor levels. The function returns a data frame with a row for
 each stratum showing the total blocks, the A-efficiency factor and the
 upper bound where available for each stratum of the design
}
\examples{
# 4 replicates of 50 treatments in complete randomized blocks
A_efficiency(blocks(treatments=50,replicates=4,blocklevels=c(4,5))$Design)
}

