% Generated by roxygen2 (4.0.2): do not edit by hand
\name{efficiencies}
\alias{efficiencies}
\title{Efficiencies of pairwise differences}
\usage{
efficiencies(Design)
}
\arguments{
\item{Design}{A block design data frame as returned by \code{\link{blocks}}}
}
\value{
\item{Incidences[[i]]}{Blocks by treatments incidence matrices  i=1... for each stratum in the design}
}
\description{
\code{efficiencies} finds efficiencies of pairwise treatment differences
}
\details{
Efficiencies of pairwise treatment differences for a design built by using the \code{blocks} function.
 The function calculates a two-way table showing the efficiency of each pairwise treatment difference.
 For equi-replicate designs only, the harmonic mean of the pairwise efficiency factors
 should equal the A-efficiency factor calculated by the \code{blocks} function.
}
\examples{
# 4 replicates of 50 treatments in complete randomized blocks

efficiencies(blocks(treatments=50,replicates=4,blocklevels=c(4,5))$Design)
}

