% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factblocks.R
\name{factblocks}
\alias{factblocks}
\title{Block designs}
\usage{
factblocks(TF, replicates = 1, rows = NULL, columns = NULL,
  order = NULL, searches = (1 + 10000\%/\%nrow(TM)), seed = sample(10000,
  1), jumps = 1)
}
\arguments{
\item{replicates}{a vector giving the replication number of each equally replicated treatment set in the \code{treatments} vector.}

\item{rows}{a vector of factor levels for the row blocks in each succesive stratum of the blocks design taken in order from the highest to the lowest.
The default is a single set of main blocks equal to the hcf of the replication numbers.}

\item{columns}{a vector of factor levels for the column blocks in each succesive stratum of the blocks design taken in order from the highest to the lowest.
The \code{rows} and the \code{columns} vectors, if both present, must be of equal length. The default is the null vector.}

\item{searches}{the maximum number of local optima searched for a design optimization. The default is 1 plus the floor of 10000 divided by the number of plots.}

\item{seed}{an integer initializing the random number generator. The default is a random seed.}

\item{jumps}{the number of pairwise random treatment swaps used to escape a local maxima. The default is a single swap.}

\item{treatments}{a vector giving a partition of the total required number of treatments into sets of equally replicated treatments.}
}
\value{
\item{Design}{Data frame giving the optimized block and treatment factors in plot order}
\item{Plan}{Data frame giving a plan view of the treatments design in the bottom stratum of the design classified by rows and columns}
\item{Efficiencies}{The achieved A- and D-efficiencies for each stratum of the design together with an A-efficiency upper-bound, where available}
\item{seed}{Numerical seed for random number generator}
\item{searches}{Maximum number of searches in each stratum}
\item{jumps}{Number of random treatment swaps to escape a local maxima}
}
\description{
Constructs randomized nested and crossed block designs for unstructured treatment sets where treatments can have any arbitrary levels of replication
and blocks can have any arbitrary feasible depth of nesting.
}
\details{
The \code{blocksdesign} package constructs arbitrary block designs with arbitrary depth of nesting and arbitrary crossed row-and-column blocks
in each nested stratum. Where one classification of a row-and-column design has a single block level, the design automatically defaults to a simple nested blocks design
(if both classifications of a stratum have a single level that stratum is dropped from the design).

 The principle design outputs comprise:
\itemize{
 \item  A data frame showing the allocation of treatments to blocks with successive nested strata arranged in standard block order. \cr
 \item  A table showing the replication number of each treatment in the design. \cr
 \item  A table showing the block levels and the achieved D- and A-efficiency factors for each blocks stratum together with A-efficiency upper bounds, where available. \cr
 \item  Plans showing the allocation of treatments to blocks or to rows and to columns in the bottom stratum of the design. \cr
}
}
\examples{

}
\references{

}

