% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_efficiencies}
\alias{pairwise_efficiencies}
\title{Pairwise efficiencies}
\usage{
pairwise_efficiencies(Design)
}
\arguments{
\item{Design}{A block design data frame as returned by \code{\link{blocks}}}
}
\value{
\item{Incidences[[i]]}{Blocks by treatments incidence matrices  i=1... for each stratum in the design}
}
\description{
\code{pairwise_efficiencies} Efficiencies of pairwise treatment differences
}
\details{
Efficiencies of pairwise treatment differences of a design built by using the \code{blocks} function.
 The output is a two-way table showing the efficiency of the treatment difference for every pair
 of treatments in a design and provides detail about the efficiency of a design for each individual
 treatment comparison. for equi-replicate designs the harmonic mean of the pairwise efficienct factors
 should equal the A-efficiency factor give by the \code{blocks} function.
}
\examples{
# 4 replicates of 50 treatments in complete randomized blocks

pairwise_efficiencies(blocks(treatments=50,replicates=4,blocklevels=c(4,5))$Design)
}

