% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_effics}
\alias{pairwise_effics}
\title{Pairwise efficiencies}
\usage{
pairwise_effics(Design)
}
\arguments{
\item{Design}{A block design data frame such as returned by \code{\link{blocks}}}
}
\description{
\code{pairwise_effics} Pairwise efficiencies of an existing design
}
\details{
Pairwise efficiencies for nested blocks design in a data frame with a
 column of block factor levels for the total blocks in each
 nested blocks stratum and a final column containing the treatment
 factor levels. The function returns a data frame with a row for
 each stratum showing the total blocks, the A-efficiency factor and the
 upper bound where available for each stratum of the design
}
\examples{
# 4 replicates of 50 treatments in complete randomized blocks

pairwise_effics(blocks(treatments=50,replicates=4,blocklevels=c(4,5))$Design)
}

