% Generated by roxygen2 (4.0.2): do not edit by hand
\name{upper_bounds}
\alias{upper_bounds}
\title{Efficiency bounds}
\usage{
upper_bounds(nplots, ntrts, nblocks)
}
\arguments{
\item{nplots}{The total number of plots in the design}

\item{ntrts}{The total number of treatments in the design}

\item{nblocks}{The total number of blocks in the design}
}
\description{
\code{upper_bounds} Finds upper A-efficiency bounds for regular block designs.
}
\details{
Upper bounds for the A-efficiency of regular nested block designs
(see Chapter 2.8 of John and Williams 1995). The treatments must have
equal replication and the block sizes must be equal therefore the number of treatments
and the number of blocks must exactly divides the number of plots.
}
\examples{
# 50 plots, 10 treatments and 10 blocks for a design with 5 replicates and blocks of size 5
upper_bounds(nplots=50,ntrts=10,nblocks=10)
}

