% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{blocks}
\alias{blocks}
\title{Block designs}
\usage{
blocks(treatments, replicates, rows = HCF(replicates), columns = NULL,
  searches = (1 + 2000\%/\%(sum(treatments) + prod(rows))),
  seed = sample(10000, 1), jumps = 1)
}
\arguments{
\item{treatments}{treatment numbers giving a partition of the total required number of treatments into sets of equally replicated treatments.}

\item{replicates}{replication numbers giving the replictaion for each set of equally replicated treatments defined by the \code{treatments} partition.}

\item{searches}{maximum number of local optima searched for a design optimization. The default is 1 plus the floor of 2000 divided by the number of model parameters.}

\item{seed}{integer initializing the random number generator. The default is a random seed.}

\item{jumps}{number of pairwise random treatment swaps used to escape a local maxima. The default is a single swap.}

\item{row_blocks}{factor levels defining the number of nested row rows in each succesive rows stratum taken in order from the highest to the lowest. 
The default is the hcf of the replication numbers.}

\item{col_blocks}{factor levels defining the number of nested column rows in each succesive rows stratum taken in order from the highest to the lowest. 
The column-rows are crossed with the row-rows in each nested stratum and the \code{row_blocks} and the \code{col_blocks} parameters, if present, must be of 
equal length and must equal the required number of nested strata.}
}
\value{
\item{Design}{Data frame giving the optimized block and treatment factors in plot order}
\item{Plan}{Data frame giving a plan view of the treatments design in the bottom stratum of the design classified by rows and columns}
\item{Efficiencies}{The achieved A- and D-efficiencies for each stratum of the design together with an A-efficiency upper-bound, where available}
\item{seed}{Numerical seed for random number generator}
\item{searches}{Maximum number of searches in each stratum}
\item{jumps}{Number of random treatment swaps to escape a local maxima}
}
\description{
Constructs randomized nested and crossed block designs for unstructured treatment sets where treatments can have any arbitrary levels of replication
and rows can have any arbitrary feasible depth of nesting.
}
\details{
The \code{treatments} and \code{replicates} parameter vectors partition the total number of treatments into sets of equally replicated treatments where 
\code{treatments} vector contains the set sizes and \code{replicates} vector contains the set replication numbers. 
The sum of the set sizes is the total number of treatments and the sum of the cross-products of the set sizes and the replication numbers
is the total number of plots. Treatments are numbered consecutively according to the ordering of the consecutive treatment sets. 

The \code{row_blocks} parameter vector, if specified, defines the nested row blocks in each individual stratum taken in order from the highest to the lowest.
The first number is the number of main row_blocks, the second, if any, is the number of nested row-blocks, the third, if any, 
is the number of nested nested row-blocks,and so on for all the reqired strata. If left blank, the default block design is a 
 maximal set of orthogonal main row-rows, where the maximal number of orthogonal rows is the highest common factor of the replication numbers. 
 
The \code{col_blocks} parameter vector, if specified, defines the nested column blocks in each individual stratum taken in order from the highest to the lowest. 
The \code{row_blocks} and \code{col_blocks} parameter vectors, if specified, must be of equal length and the pairs of parameters for each
particular stratum defines the row-and-column block design for that stratum. 
If a simple nested blocks design is required for any particular stratum then the number of columns for that stratum should be set to one and the number of rows to
the required number of nested blocks. If the \code{col_blocks} parameter vector is left null, a simple nested blocks design defined by the \code{row_blocks} parameter
is fitted.

The block sizes in each rows stratum including the row-by-columns interactions strata are always as equal as possible and never differ in size by more than a single unit
for replicated treatments. 
If the number of nested rows in a particular stratum exactly divdes the number of units, the block sizes in that stratum will be exactly equal; otherwise the block sizes
will differ by at most one unit. Row blocks and column blocks must always comprise at least two plots while the row-column intersection rows may comprise a single plot
or any feasible number of plots. 

Unreplicated treatments with a single replication can also be included but are added heuristcally by assigning single unreplicated treatments to the optimized blocks of the
replicated part of the design. To ensure that the blocks of an augmented design containing unreplicated treatments are as near equal as possible, the optimized blocks of the
replicated part of the design must all be equal in size in any particular stratum of the design.     

 Warning: The swapping algorithm works from the top stratum downwards and is always constrained to make improving swaps within the levels of any existing rows.
 This can be problematic for a 2-replicate treatment design arranged as a 2 x 2 row-and-clolumn design because the algorithm will find a semi-Latin squarere with complete
 replicate row-blocks and complete replicate column-blocks which means that one treatment contrast will be
 confounded with the row-by-column interaction effect and the algorithm will be unable to fit any nested block design within the rows of 
 any 2 x 2 semi-Latin square. For this situation, it would be better to construct a simple block design with four main row-blocks and a single main column-block 
 and then to nest any required sub-block design within the four main blocks.   
 
 Lattice designs where v is a prime-power require the \code{\link[crossdes]{MOLS}} package.

 The principle design outputs comprise:
\itemize{
 \item  A design matrix showing the allocation of treatments to rows with successive nested rows factors arranged in successive columns in standard block order.  \cr
 \item  A design matrix as above but with the last (bottom) rows factor shown arranged horizontally to give a plan view. \cr
 \item  A set of incidence matrices, one for each rows stratum, showing the number of times each treatment occurs in each block for each stratum. \cr
 \item  A table showing the achieved D- and A-efficiency factors for each nested rows stratum together with an A-efficiency upper bound, where available. \cr
 \item  A table showing a skeleton analysis of degrees of freedom for the combined block and treatment design. \cr
} 

Very occasionally, the algorithm may fail to converge due to a near-singular design with a large number of single plot rows.
In that case, it may be best to build a simpler design with larger rows and then to add the extra block constraints by hand using ad hoc or heuristic methods.
}
\examples{

# 3 treatments x 2 replicates, 2 treatments x 4 replicates and 4 treatments x 3 replicates  
# the hcf of the replication numbers is 1 therefore the default design is completely randomized 
rows(treatments=c(3,2,4),replicates=c(2,4,3))

# 4 treatments x 4 replicates with 2 main rows each containing two complete replicates  
rows(treatments=4,replicates=4,blocklevel=2)

# 50 treatments x 4 replicates with 4 main rows and 5 nested sub-rows in each main block 
rows(treatments=50,replicates=4,rows=c(4,5))

# as above but with 20 additional single replicate treatments 
# giving exactly one single replicate treatment per sub-block
rows(treatments=c(50,20),replicates=c(4,1),rows=c(4,5))

# 64 treatments x 2 replicates with 2 main rows and five succesively nested 2-level factors
rows(treatments=64,replicates=2,rows=c(2,2,2,2,2,2))

# 6 replicates of 6 treatments in 4 rows of size 9 (non-binary block design)
rows(treatments=6,replicates=6,rows=4)

# concurrence matrix of balanced incomplete block design 
crossprod(rows(13,4,13,searches=100)$Incidences[[1]])

# concurrence matrix for 13 treatments x 4 replicates and 13 treatments with one rep in 13 rows 
crossprod(rows(c(13,13),c(4,1),13)$Incidences[[1]])

# 2**10 treatments x 2 replicates in 2**10 rows giving a fully saturated rows design 
# (requires a considerable time to run!)
\dontrun{ d=rows(1024,2,rep(2,10)) }
         
}
\references{
Sailer, M. O. (2013). crossdes: Construction of Crossover Designs. R package version 1.1-1. http://CRAN.R-project.org/package=crossdes
}

