% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blocks}
\alias{blocks}
\title{Block designs}
\usage{
blocks(treatments, replicates, blocklevels = hcf, searches = min(64,
  floor(4096/nunits)), seed = NULL)
}
\arguments{
\item{treatments}{A list of the treatment numbers for each treatment set of the design. The list should contain one set of treatment numbers
for each replication level in the design. Treatments are numbered consecutively according to the ordering of the treatment sets.}

\item{replicates}{A list of the replication numbers for each treatment set in the design.
The treatments and replications lists must be equal in length and must contain matching pair of numbers, one pair for each treatment replication set.}

\item{blocklevels}{An optional list of nested blocks.
The numbers in the list define the number of nested blocks in each blocks stratum with the first number
giving the number of main blocks and the succesive numbers, if any, giving the number of nested
blocks in each preceding block. The default is the hcf of the replication numbers.}

\item{searches}{An optional parameter for the number of local optima searched during
a design optimization. The default gives a reasonably short search time in most situations.}

\item{seed}{An optional integer seed for initializing the random number generator}
}
\value{
\item{Design}{Data frame showing block and treatment factors for each plot}
\item{Plan}{Data frame showing treatments allocation to plots for each block}
\item{Incidences[[i]]}{Blocks-by-treatments incidence matrix for each stratum in the design where (i = 1... strata) }
\item{Efficiencies}{Data frame showing the A-efficiency and an upper bound, where available, for each stratum in the design}
}
\description{
\code{blocks} constructs nested block designs for unstructured treatment sets with arbitrary replication and arbitrary depth of nesting
}
\details{
\code{blocks} constructs nested block designs for unstructured treatment sets where treatments can have any arbitrary replication
and blocks can be nested to any feasible depth of nesting.

Treatment and replication numbers are defined by \code{treatments} and \code{replicates} parameter lists where matching pairs of numbers in the two lists
represent sets of treatments with that number of treatments and that level of replication. Any number of treatment sets are allowed
and treatments are  numbered consecutively according to the ordering of the treatment sets in the parameter lists (see the first example).

Block sizes in the same stratum will be equal whenever possible and will never differ by more than a single unit.
Blocks strata are optimized hierarchically with the blocks of each new set optimized within the blocks of the preceding set.

\code{blocks} optimizes general designs algorithmically by a swapping algorithm that maximizes the determinant of
the information matrix (D-optimality). However, certain special lattice designs (see vignette) are constructed algebraically
using mutually orthogonal latin squares (MOLS).  Lattice designs with a number of treatments equal to the square of a prime-power
and with more than three replicates are constructed by using the MOLS function of the \code{crossdes} package.

Treatment designs are fully randomized with treatments randomized within the bottom blocks of the design and each set of nested blocks
randomized within each preceding set of blocks.
}
\examples{
# 2 reps of 3 treatments, 4 reps of 2 treatments and 3 reps of 4 treatments with hcf(2,4,3) = 1
blocks(treatments=c(3,2,4),replicates=c(2,4,3))

# 4 replicates of 50 treatments in 4 complete randomized blocks where hcf(4) = 4
blocks(treatments=50,replicates=4)

# as above but with 4 main blocks and 5 nested blocks in each main block
blocks(treatments=50,replicates=4,blocklevels=c(4,5))

# as above but with 20 additional single replicate treatments
blocks(treatments=c(50,20),replicates=c(4,1),blocklevels=c(4,5))

# tabulation of 4 reps of 4 treatments and 8 reps of 1 treatment with 8 nested blocks per main block
table(blocks(treatments=c(4,1),replicates=c(4,8))$Design)

# concurrences for 3 replicates of 36 treatments with 3 main blocks and 6 nested blocks
crossprod(blocks(treatments=36,replicates=3,blocklevels=c(3,6))$Incidences[[2]])

# 4 replicates of 64 treatments with 4 main blocks and five 2-level nesting factors
blocks(treatments=64,replicates=4,blocklevels=c(4,2,2,2,2,2))

# concurrences for 4 reps of 13 treatments in 13 blocks with 13 added single rep treatments
crossprod(blocks(c(13,13),c(4,1),13,searches=100)$Incidences[[1]])
}

