% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blocks}
\alias{blocks}
\title{Block designs}
\usage{
blocks(treatments, replicates, blocklevels = NULL, searches = NULL,
  seed = NULL)
}
\arguments{
\item{treatments}{a list partitioning the total number of treatments into equally replicated treatment sets.}

\item{replicates}{a list assigning a replication level to each equally replicated treatment set.}

\item{blocklevels}{a list of nested block levels where the first level is the number of main blocks
and the remaining levels, if any, are the levels of sub-blocks in  a hierarchy of nested sub-blocks.
The default setting is the highest common factor of the \code{replicates} list.}

\item{searches}{the number of local optima searched during optimization. The default setting is the minimum of 64 or the integer quotient of 4096
divided by the number of plots.}

\item{seed}{an integer seed for initializing the random number generator where a design must be reproducible. The default
setting is a random seed.}
}
\value{
\item{Design}{Data frame showing the listing of treatments allocated to blocks}
\item{Plan}{Data frame showing a plan of treatments allocated to sub-plots within blocks}
\item{Incidences}{List of blocks-by-treatments incidence matrices, one for each stratum in the design}
\item{Efficiencies}{Data frame showing the A-efficiency factor for each stratum in the design together with an upper bound, where available}
\item{seed}{Numerical seed for random number generator}
}
\description{
Constructs nested block designs for unstructured treatment sets with arbitrary replication, not necessarily all equal, and arbitrary depth of nesting.
}
\details{
The \code{blocks} function constructs optimized nested block designs for unstructured treatment sets where treatments can have any arbitrary replication, not necessarily all equal,
and blocks can have any feasible depth of nesting.

The \code{treatments} list contains the total number of treatments in the design partitioned into
sets of equally replicated treatments where the replication can vary between sets but must be constant within sets.
The \code{replicates} list provides the replication level for each treatment set and must be the same length as
the \code{treatments} list. The individual treatments are labelled consecutively according to the ordering of the sets
in the two lists and different treatment sets with the same replication can be used, if required.

The \code{blocklevels} list defines the blocks structure of the design where the first level in the list is the number of main blocks
and the succesive levels, if any, are the levels of succesive sub-blocks in a hierarchy of sub-blocks. The length of the list is the number of strata
and the running products of the levels are the total numbers of
blocks in the succesive strata. The average block size of a stratum is the total number of blocks divided by
the total number of plots andthe block sizes are all equal if the quotient is an integer. Otherwise block sizes differ by, at most, a single unit.
The default design is a complete set of othogonal main blocks.

General block designs are constructed algorithmically by a swapping algorithm that seeks to maximize the determinant of
the information matrix (D-optimality). Beginning with the main blocks stratum, pairs of treatments are swapped at random between
sub-blocks nested within existing blocks until a local optima is attained. If the number of searches
is greater than one, the algorithm then makes a number of random swaps and then continues with improving swaps
until another local optima is reached. After the required number of searches, the best overall design
is restored and the process repeated for the next stratum in the hierarchy. Eventually the bottom stratum
is reached after which the algorithm stops. Special lattice designs for v**2 equally replicated treatments in blocks of size v with k replicates
are constructed algebraicaly when k <= 3 or when v is prime or prime-power and k <= v+1 or when v = 10 and k <= 4.
The \code{crossdes} package is used for lattice designs with prime-power v.

Optimized designs are fully randomized with each set of nested blocks randomized within the preceding set of blocks and with
treatments randomized within the bottom set of blocks.
}
\examples{
# 3 treatments with 2 reps, 2 treatments with 4 reps, 4 treatments with 3 reps
# the hcf of the replication numbers is 1 and the default design is a completely randomized design
blocks(treatments=c(3,2,4),replicates=c(2,4,3))

# 50 treatments with 4 reps in 4 complete randomized blocks
blocks(treatments=50,replicates=4)

# as above but with 4 main blocks and 5 nested blocks within each main block
blocks(treatments=50,replicates=4,blocklevels=c(4,5))

# as above but with 20 additional single replicate treatments, one to each block
blocks(treatments=c(50,20),replicates=c(4,1),blocklevels=c(4,5))

# 64 treatments with 2 reps and 2 main blocks with five 2-level nested factors
blocks(treatments=64,replicates=2,blocklevels=c(2,2,2,2,2,2),searches=4)

# concurrence matrices of 36 treatments with 3 reps and 3 main blocks with 6 nested blocks
crossprod(blocks(treatments=36,replicates=3,blocklevels=c(3,6))$Incidences[[2]])

# concurrence matrix for 13 treatments with 4 reps and 13 treatments with one rep in 13 blocks
crossprod(blocks(c(13,13),c(4,1),13,searches=100)$Incidences[[1]])
}
\references{
Sailer, M. O. (2013). crossdes: Construction of Crossover Designs. R package version 1.1-1. http://CRAN.R-project.org/package=crossdes
}

