% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blocks}
\alias{blocks}
\title{Block designs}
\usage{
blocks(treatments, replicates, blocklevels = NULL, searches = NULL,
  seed = NULL)
}
\arguments{
\item{treatments}{a list giving a partition of the total number of treatments into equally replicated treatment sets.}

\item{replicates}{a list assigning a replication level to each set in the \code{treatments} list.}

\item{blocklevels}{a list of levels where the first level is the number of main blocks and the remaining
levels, if any, are the levels of a hierarchy of nested sub-blocks in descending order of nesting.
The default is an orthogonal main blocks design.}

\item{searches}{the number of local optima searched during an optimization. The default is the minimum of 64 or
 4096 divided by the number of units.}

\item{seed}{an integer seed for initializing the random number generator. The default
is a random seed.}
}
\value{
\item{Design}{Data frame showing the listing of treatments allocated to blocks}
\item{Plan}{Data frame showing a plan of treatments allocated to sub-plots within blocks}
\item{Incidences}{List of blocks-by-treatments incidence matrices, one for each stratum of the design}
\item{Efficiencies}{Data frame showing the A-efficiency factor for each stratum of the design together with an upper bound, where available}
\item{seed}{Numerical seed for random number generator}
}
\description{
Constructs nested block designs for unstructured treatment sets with arbitrary replication,
not necessarily all equal, and arbitrary depth of nesting.
}
\details{
\code{blocks} constructs general block designs that aim to maximize the determinant of
the information matrix (D-optimality). The general design algorithm works by making improving swaps in the top stratum
of a design until no further improvement is possible and then repeating
the process for each nested stratum in turn until the bottom stratum is reached.
At each stage, improving swaps are nested within existing blocks to ensure top-down optimization.

Certain special lattice designs with v**2 equally replicated treatments in blocks of size v and with k replicates
are optimized algebraically for k <= 3 or for v prime or prime-power and k <= v+1 or for v = 10 and k <= 4.
\code{crossdes} is used for lattice designs with prime-power v.


\code{treatments} is a parameter list that partitions the total number of treatments into
sets of equally replicated treatments. All treatments in the same set
have equal replication and all treatments are numbered consecutively according to their set order.
Normally, all treatments with the same replication will occur in the same set but
treatments with the same replication could be split between two or more sets  if, say, a non-standard treatment ordering
 was required.

\code{replicates} is a parameter list of replication levels for the sets in the treatments list.
  The treatments and replicates lists must be of equal length.

\code{blocklevels} is a parameter list of levels for the blocks design where the first number is the number of main blocks
and the succesive numbers, if any, are the numbers of sub-block levels in a hierarchy of nested sub-blocks.
The length of the list is the number of strata and the running products are the total numbers of
blocks in each succesive stratum. The blocks in any given stratum are all equal in size or differ by, at most, a
single plot. The default setting is the highest common factor of the replication levels which gives a main blocks
 design with a maximal set of complete othogonal main blocks.

\code{searches} is the number of local optima searched for each stratum for during an optimization with the
 best optima selected before proceeding to the next stratum. Each new search from a local optima makes a number of
 random swaps (6) to escape the optima and then proceeds by improving swaps. The number of searches can be set to
 any value, as required.

Optimized designs are fully randomized with each set of nested blocks fully randomized within preceding sets of blocks and with
treatments fully randomized within blocks.
}
\examples{
# 3 treatments with 2 reps, 2 treatments with 4 reps, 4 treatments with 3 reps
# the hcf of the replication numbers is 1 and the default design is a completely randomized design
blocks(treatments=c(3,2,4),replicates=c(2,4,3))

# 50 treatments with 4 reps in 4 complete randomized blocks
blocks(treatments=50,replicates=4)

# as above but with 4 main blocks and 5 nested blocks within each main block
blocks(treatments=50,replicates=4,blocklevels=c(4,5))

# as above but with 20 additional single replicate treatments, one to each block
blocks(treatments=c(50,20),replicates=c(4,1),blocklevels=c(4,5))

# 64 treatments with 2 reps and 2 main blocks with five 2-level nested factors
blocks(treatments=64,replicates=2,blocklevels=c(2,2,2,2,2,2),searches=4)

# concurrence matrices of 36 treatments with 3 reps and 3 main blocks with 6 nested blocks
crossprod(blocks(treatments=36,replicates=3,blocklevels=c(3,6))$Incidences[[2]])

# concurrence matrix for 13 treatments with 4 reps and 13 treatments with one rep in 13 blocks
crossprod(blocks(c(13,13),c(4,1),13,searches=100)$Incidences[[1]])
}
\references{
Sailer, M. O. (2013). crossdes: Construction of Crossover Designs. R package version 1.1-1. http://CRAN.R-project.org/package=crossdes
}

