% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blocks}
\alias{blocks}
\title{Block designs}
\usage{
blocks(treatments, replicates, blocklevels = NULL, searches = NULL,
  seed = NULL)
}
\arguments{
\item{treatments}{a list giving a partition of the total number of treatments into
sets where all treatments in the same set have the same replication.}

\item{replicates}{a list assigning the replication level of each set in the \code{treatments} list.}

\item{blocklevels}{a list of block levels where the first level is the number of main blocks and the remaining
levels, if any, are the succesive nested levels of a hierarchy of nested sub-blocks.
The default is an orthogonal main blocks design.}

\item{searches}{an optional integer for the number of local optima searched during an optimization.
The default is the minimum of 32 or the ceiling of 4096 divided by the number of units.}

\item{seed}{an optional integer seed for initializing the random number generator. The default
is a random seed.}
}
\value{
\item{Design}{Data frame showing the listing of treatments allocated to blocks}
\item{Plan}{Data frame showing a plan of treatments allocated to sub-plots within blocks}
\item{Incidences}{List of blocks-by-treatments incidence matrices, one for each stratum of the design}
\item{Efficiencies}{Data frame showing the A-efficiency factor for each stratum of the design together with an upper bound, where available}
\item{seed}{Numerical seed for random number generator}
}
\description{
Constructs nested block designs for unstructured treatments with arbitrary replication,
not necessarily all equal, and arbitrary depth of nesting.
}
\details{
\code{blocks(...)} provides top-down optimization of the succesive blocks strata of a nested blocks design. The
 blocks of each nested stratum are optimized within the blocks of the containing stratum so the
 blocks of any contained strata are conditional on the blocks of the containing strata whereas the blocks of any
 containing strata are independent of the blocks of the contained strata.

If the blocks in the top stratum have k replicates with v**2 equally replicated treatments in blocks of size v
and k <= 3 for any v, or k <= v+1 for prime or prime-power v, or k <= 4 for v = 10,
then the stratum is a lattice block design and is constructed algebraically.

All other blocks are constructed algoritmically by a D-optimality algorithm that makes
improving swaps between nested blocks within containing blocks until no further improvement is possible.

\code{treatments} is a list of sets where the sum of the sets is the required number of treatments
and the treatments in any one set are all equally replicated.

\code{replicates} is a list of replication numbers for sets in the \code{treatments} list.
Treatments are numbered consecutively according to the order of the sets
and treatments with the same replication can be split between two or more sets if required.

\code{blocklevels} is a list of nested block levels for the succesive nested blocks strata of the design.
The first level is the number of main blocks
and the successive levels, if any, are the numbers of sub-blocks in the succesive strata of
the nested blocks design.
The length of the list is the number of strata  and the
running products of the levels are the total blocks in each successive stratum of the
design. Blocks in the same stratum are always equal in size or differ by, at most, a
single unit. The default is the highest common factor of the replication levels,
which gives a main blocks design with a maximal set of complete orthogonal main blocks.

The \code{searches} parameter is the number of local optima searched during an optimization.
Increasing the number of searches may improve the efficiency of a design but
 will also increase the search time.

The \code{seed} parameter is an integer used to initialize the random number generator. The
 default is a random integer but any fixed positive integer can be used instead,if required.

Blocks and treatments are fully randomized within the constraints of a nested blocks design.
}
\examples{
# 3 treatments with 2 reps, 2 treatments with 4 reps, 4 treatments with 3 reps
# the hcf of the replication numbers is 1 and the default design is a completely randomized design
blocks(treatments=c(3,2,4),replicates=c(2,4,3))

# 50 treatments with 4 reps in 4 complete randomized blocks
blocks(treatments=50,replicates=4)

# as above but with 4 main blocks and 5 nested blocks within each main block
blocks(treatments=50,replicates=4,blocklevels=c(4,5))

# as above but with 20 additional single replicate treatments, one to each block
blocks(treatments=c(50,20),replicates=c(4,1),blocklevels=c(4,5))

# 64 treatments with 2 reps and 2 main blocks with five 2-level nested factors
blocks(treatments=64,replicates=2,blocklevels=c(2,2,2,2,2,2),searches=4)

# concurrence matrices of 36 treatments with 3 reps and 3 main blocks with 6 nested blocks
crossprod(blocks(treatments=36,replicates=3,blocklevels=c(3,6))$Incidences[[2]])

# concurrence matrix for 13 treatments with 4 reps and 13 treatments with one rep in 13 blocks
crossprod(blocks(c(13,13),c(4,1),13,searches=100)$Incidences[[1]])

#
blocks(treatments=4,replicates=3,blocklevels=c(3,2))
blocks(treatments=4,replicates=3,blocklevels=c(3,2))
 blocks(treatments=4,replicates=3,blocklevels=c(3,2))
blocks(treatments=4,replicates=3,blocklevels=c(3,2))
blocks(treatments=4,replicates=3,blocklevels=c(3,2))
blocks(treatments=4,replicates=3,blocklevels=c(3,2))
 blocks(treatments=4,replicates=3,blocklevels=c(3,2))
 blocks(treatments=4,replicates=3,blocklevels=c(3,2))
 blocks(treatments=4,replicates=3,blocklevels=c(3,2))
}
\references{
Sailer, M. O. (2013). crossdes: Construction of Crossover Designs. R package version 1.1-1. http://CRAN.R-project.org/package=crossdes
}

