% Generated by roxygen2 (4.0.2): do not edit by hand
\name{efficiencies}
\alias{efficiencies}
\title{Efficiencies of pairwise differences}
\usage{
efficiencies(Design)
}
\arguments{
\item{Design}{A block design data frame as returned by \code{\link{blocks}}}
}
\value{
\item{Incidences[[i]]}{Blocks by treatments incidence matrices  i=1... for each stratum in the design}
}
\description{
\code{efficiencies} tabulates the efficiency of each pairwise treatment difference for each stratum of a design
}
\details{
Tabulates the efficiency of each pairwise treatment difference for each statum of a design built by the \code{blocks} function.
 The table can be extracted and used for comparison of different designs. For example with equi-replicate designs,
 the harmonic mean of the pairwise efficiency factors
 will equal the A-efficiency factor calculated by the \code{blocks} function.
}
\examples{
# 4 replicates of 50 treatments in complete randomized blocks
efficiencies(blocks(treatments=50,replicates=4,blocklevels=c(4,5))$Design)
}

